Add /operations endpoint to list available operations
This is a straightforward change to add a `/operations` endpoint which
can be used to get a list of the available operations on the CyberChef
server. The returned object has an attribute for each operation name,
with the argument description as its value.

This change also documents this new endpoint in `README.md` and
`swagger.yml`.